requirements.txt
sqlalchemy>=1.3.20,<1.4
alembic>=1.4.3,<1.5
flask>=1.1.2,<1.2
flask-bootstrap4>=4.0.2,<4.1
flask-sqlalchemy>=2.4.4,<2.5
flask-migrate>=2.5.3,<2.6
flask-wtf>=0.14.3,<0.15
psycopg2-binary>=2.8.6,2.9
